Cleaner implementation
