+++ pyproject.toml
@@ -22,8 +22,7 @@ license = {text = "GNU General Public License v3"}
 dependencies = [
   "django",
   "django-extensions",
-  "ipython",
-  "python-dotenv"
+  "ipython"
 ]
 
 [project.optional-dependencies]
@@ -31,7 +30,6 @@ test = [
     "django",
     "django-extensions",
     "ipython",  
-    "python-dotenv",
     "black",  # code auto-formatting
     "coverage",  # testing
     "isort",  # code auto-formatting
